refactor: change hook placement
