+++ python/hedera_agent_kit/__init__.py
@@ -6,6 +6,15 @@ __all__ = [
     "Plugin",
     "HederaMCPServer",
     "Context",
+    "AbstractHook",
+    "Policy",
+    "HcsAuditTrailHook",
+    "MaxRecipientsPolicy",
+    "RejectToolPolicy",
+    "PostCoreActionParams",
+    "PostParamsNormalizationParams",
+    "PostSecondaryActionParams",
+    "PreToolExecutionParams",
 ]
 
 # Re-export key SDK primitives from the shared package
@@ -17,4 +26,16 @@ from .shared import (
     Plugin,
     HederaMCPServer,
     Context,
+    AbstractHook,
+    Policy,
+    PostCoreActionParams,
+    PostParamsNormalizationParams,
+    PostSecondaryActionParams,
+    PreToolExecutionParams,
 )
+
+# Re-export concrete hooks
+from .hooks import HcsAuditTrailHook
+
+# Re-export concrete policies
+from .policies import MaxRecipientsPolicy, RejectToolPolicy
