+++ python/hedera_agent_kit/hooks/__init__.py
@@ -1,0 +1,3 @@
+from .hcs_audit_trail_hook import HcsAuditTrailHook
+
+__all__ = ["HcsAuditTrailHook"]
