+++ python/hedera_agent_kit/policies/__init__.py
@@ -1,3 +1,4 @@
-__all__ = ["MaxRecipientsPolicy"]
+__all__ = ["MaxRecipientsPolicy", "RejectToolPolicy"]
 
-from hedera_agent_kit.policies.max_recipients_policy import MaxRecipientsPolicy
+from .max_recipients_policy import MaxRecipientsPolicy
+from .reject_tool_policy import RejectToolPolicy
