+++ python/hedera_agent_kit/shared/__init__.py
@@ -7,10 +7,24 @@ __all__ = [
     "Plugin",
     "HederaMCPServer",
     "Context",
+    "AbstractHook",
+    "Policy",
+    "PostCoreActionParams",
+    "PostParamsNormalizationParams",
+    "PostSecondaryActionParams",
+    "PreToolExecutionParams",
 ]
 
+from .abstract_hook import (
+    AbstractHook,
+    PostCoreActionParams,
+    PostParamsNormalizationParams,
+    PostSecondaryActionParams,
+    PreToolExecutionParams,
+)
 from .api import HederaAgentAPI
 from .configuration import Configuration, AgentMode, HederaMCPServer, Context
 from .plugin import Plugin
+from .policy import Policy
 from .tool import Tool
 from .tool_discovery import ToolDiscovery
